+++ glibc/ctype/isspace.c
@@ -1,0 +1,16 @@
+#include <stdio.h>
+#include <ctype.h>
+
+int main(int argc, char *argv[])
+{
+#define S(ch) printf("%-4s is%s space\n", #ch, isspace(ch) ? "" : " not");
+
+	S(' ');
+	S('\t');
+	S('\r');
+	S('\n');
+	S('\b');
+	S('a');
+
+	return 0;
+}